test: integrate handlebars and boost.url test suite
This change integrates the test suite library developed for handlebars with the test suite library from boost.url that is being used for other tests.

With this integration, Handlebars tests are listed among any other tests in the library.

The features used for handlebars, such as the expression decomposer and the diff algorithm for golden master tests, have been integrated into the test library from Boost.URL.

fix #390, #384