test: JavaScript unit tests
Implements unit tests for the Javascript bindings based on duktape. All errors found are fixed in the same commit.

This commit also includes the complete implementation for classes that were only placeholders, including conversions to dom::Value and related wrapper classes for objects, arrays, and functions.

The changes are only related to bindings from Javascript to C++. Bindings from C++ to Javascript will be addressed in a later commit.