ci: statically link libc
Move configuration values to the matrix action and statically links both `libgcc` and `libstdc++`.

Since we are using the latest version of ubuntu in our workflow, this is intended to allow the release binaries to work on any earlier version of ubuntu. This is required by the Antora scripts in the Boost release tools.