ci: create release for Ubuntu 20.04
This commit adds a new workflow that creates a Linux release on Ubuntu 20.04. This allows the release package to be used on more versions of Ubuntu. Since the C++ compilers we need are not available in Ubuntu 20.04, we use the clang executable from our own LLVM built from source to build the dependencies and MrDocs.