refactor: simplify handlebars generators
Apart from the DocVisitor, the Adoc and HTML generators are nearly 100% duplicated code.

This is the main reason the HTML Generator is always behind the Adoc Generator: simple features were not duplicated in the HTML generator. This even justified #656, where we would temporarily remove the HTML generator.

This commit unifies the standard code to remove duplicated code, simplify maintenance, and ensure that both generators always have access to the same features.

In passing, many missing features in the HTML generator are now available.

fix #656