refactor: reimplement ADT/Polymorphic
This is a simpler design, with less differences in API wrt to
std::polymorphic, though still with some differences, includinding
nullability, and some unused features which are not important to us
are left out.

This avoids memory leaks facilitated by the previous design, and is
a step towards making the project pass the LeakSanitizer.