feat: AsciiDoc and HTML outputs support the final specifier for classes and class templates
This was already supported in XML. In this commit, we also remove the
semicolon when that would produce invalid C++ code, e.g. in

class A final;

or

class A : B1, B2;

Instead, "{ /* see-below */ }" is now always followed by a semicolon,
even if the record is final and/or base classes are present.

This closes issue #963.