feat: use system libs by default
This change makes mrdocs use the system includes by default instead of
the bundled ones.

This helps mrdocs build projects which rely on differences between
standard library implementations, since now mrdocs sees the source code
closer to what the original compiler did, with only potential differences in the
compilers themselves and some minor command line manipulations remaining.

It also stops bundling the clang resource directory, relying on the one
installed with libclang. This makes it so installing both to the same
prefix be necessary, which is the same as all the other libclang based tools.