don't mark unnamed classes as implementation defined
These classes can appear as user written code and needs to be
documented, as show in the new test case.

In general, for implementation defined declarations, we can detect them
as the clang AST exposes them as 'implicit', but this is possible
for any declaration, and can be handled more generally.

The new included exposes a different bug, where we don't attach the
documentation provided for each member in a declarator, which clang
does expose in the AST, but we seem to be ignoring it.

Fixes #1035